requirements.txt
astroid==1.6.5
certifi==2018.10.15
chardet==3.0.4
Click==7.0
decorator==4.3.0
diff-match-patch==20121119
Django==2.1
django-import-export==1.1.0
django-searchable-select==1.5.0
et-xmlfile==1.0.1
future==0.17.1
geocoder==1.38.1
idna==2.7
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.4
odfpy==1.3.6
openpyxl==2.5.8
pandas==0.23.4
postgis==1.0.4
psycopg2==2.7.5
pylint==1.9.3
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
ratelim==0.1.6
requests==2.20.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
